dird: moving reload utilities to their own files
