+++ core/src/dird/dird.h
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2008 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -217,8 +217,6 @@ struct runtime_job_status_t {
 #define FD_VERSION_52 52
 #define FD_VERSION_53 53
 #define FD_VERSION_54 54
-
-bool DoReloadConfig();
 
 } /* namespace directordaemon */
